[TENSORS] Made the 'TensorLike' trait sealed.
